component splitup
